AI_BASED_QUESTIONER/requirements.txt
setuptools
pycryptodome
dnspython
Flask
firebase-admin
pyrebase4
gcloud  
ipython
openai-whisper  
pytube
pypdf
google-generativeai 
google


# pip install setuptools

# pip install pycryptodome

# pip install Flask

# pip install firebase_admin

# pip install pyrebase4

# pip3 install --upgrade gcloud

# pip install ipython

# pip install -U openai-whisper

# pip install git+https://github.com/openai/whisper.git

# pip install pytube/python -m pip install git+https://github.com/pytube/pytube

# pip install pypdf

# pip install -U google-generativeai

# pip install google
